% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_best_t_for_cell}
\alias{get_best_t_for_cell}
\title{Find best T for a single cell}
\usage{
get_best_t_for_cell(cell_vec, B, BR, R, omega1, omega2)
}
\description{
Find best T for a single cell
}
\keyword{internal}
